Fixed monthly gb sensors and usage sensors set to 0 at integration loading
When the integration is starting or the polling of data statuses fails, the sensor was writing 0.0 which caused fatal issues in the statistic. This is now fixed, so when the integration reads 0.0 it's set to "unavailable".

This does not work for rx_bytes and tx_byes, because these sensors are automatically created. But as nobody like to read bytes (we have a GB sensor for better reading which integrates the 0.0 fix) we can disregard it.